Completing week02 group activity. Finished Step 6.
